Converter from dat to info(meta) structure
dat2info.m will extract information from the dat structure (which is the
output o the GUI) to the old format info/meta structure.
Somebody please update for the multichannel cases.
dat2infoWrapper will do the loading and saving as well, but be careful
not to overwrite your data (the script is NOT checking if the file
already exists)
Use carefully, might still be buggy, nut worked for me.